Move fee logic to `zksync_types`
